+++ src/detection/temps/temps_nvidia.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <stdint.h>
+
+const char* ffDetectNvidiaGpuTemp(double* temp, uint32_t deviceId);